+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [val] P7MValidator in full validator facade
 - [playground-quarkus] P7MValidator in validator feature
 
+### Changed
+
+- [playground-quarkus] docker image eclipse-temurin:21.0.1_12-jre-alpine
+
 ## [3.2.0] - 2023-11-02
 
 ### Added